Switch KDoc formatting to use KDoc Parser from Kotlin suite
Summary:
This deletes the cloned JavaDoc Lexer, and uses the one that comes with the Kotlin compiler suite.
I am parsing and converting the tokens in the minimal (and somewhat ugly) way that is needed to get reasonable tests passing.
From here we can clean it up, and support more formats in KDoc properly.

Reviewed By: cgrushko

Differential Revision: D19626397

fbshipit-source-id: b4b33e450bff8aac95c9bd4eabfc04216285774d