Redo KDoc lexing and parsing so it's simpler to reason about
Summary:
This rewrites more of the code so that we do a simpler lexical transform, and move some of the code to the writer to try and resemble the original code's intent of style.
This will make it easier to support the missing tokens for now in the future diffs.

Reviewed By: cgrushko

Differential Revision: D19648752

fbshipit-source-id: 96f1260e3707340cef27edecefb2f656263ce8ec