Format using replacements instead of dumping the entire formatted output
Summary:
This is the first step for formatting ranges, which is nice to avoid git/hg-blame churn when changing the formatter.
In order for this to work properly, we need to pepper KotlinInputAstVisitor with calls to `markForPartialFormat();`, to tell the formatter what constitutes the smallest formattable range. This is left for a follow-up diff.

Reviewed By: strulovich

Differential Revision: D19843183

fbshipit-source-id: dcd723de07e528223b94f430273d4f35a3fda262