Treat chain expressions with one lambda in the end differently
Summary:
This has been driving me (And others) for a while. IT just didn't feel right for many cases, like the let example in the test here.

The solution is to treat chains with exactly one lambda in their end in a special way where we try to keep everything up to the first `{` in the same line. This cleans up quite a few awkward things.

We avoid this new behavior if the lambda is not the final expressions to avoid this weird situation:
```
foo.bar.bar?.let {
  // ...
}
    .nextAction()
```

And we don't want this as an alternative:
```
foo.bar.bar?.let {
  // ...
}.nextAction()
```
(Since it doesn't fit out rules well)

Reviewed By: cgrushko

Differential Revision: D20258364

fbshipit-source-id: d7287a7f13fb33e1c464ad24128543070c540131