In function calls with single lambda in parenthesis, keep parenthesis and braces together
Summary:
When a call with only one arg and a lambda is being made in parenthesis, it looks nicer if we don't add spaces between the braces and the parenthesis.

One may wonder - why would anyone use this style? There are a few reason, but the main one is calling constructors which take lambdas.

Reviewed By: cgrushko

Differential Revision: D20363931

fbshipit-source-id: fffc087b9ffea5211f8829c3208302cd62f973e2