Sort imports
Summary:
Anything but imports and whitespace is not allowed between imports, and causes a parsing error. In particular, this makes ktfmt not support comments between import directives (comments after are fine).

This is done by getting the imports, sorting them, and then overwriting the original code string in the appropriate area. This means that all whitespace is lost even before it arrives at the formatter, but that's fine, because the formatter ignores whitespace anyway.

Partially addresses Issue 17.

Reviewed By: strulovich

Differential Revision: D20583202

fbshipit-source-id: f9c7ae2b48ca6b6a480fa514791ac79134cacbc7