Handle implicit and malformed code blocks by adding explicit markers
Summary:
A bunch of issues throw off the KDoc formatter when code blocks happen with three-backticks in the middle of the line, or also code blocks with no three-backticks.

This normalizes our handling of it, and adds ``` into code blocks when the lexer decided a token is a code block, but we did not see an explicit marker.

Reviewed By: cgrushko

Differential Revision: D20515472

fbshipit-source-id: a7dceeaccc249f7bceda488bb32991d1757e98ab