Preserve trailing spaces in multiline strings
Summary: google-java-format removes trailing spaces when it emits formatted code, which is a problem for multiline string literals. We trick it by replacing the last trailing space in such cases with a tombstone, a character that's unlikely to be used in a regular program. After formatting, we replace it back to a space.

Reviewed By: strulovich

Differential Revision: D20647586

fbshipit-source-id: c7e7ed9658a81638060e05f2c0360cff68bc8d69