Redo qualified expressions in a safer way
Summary:
For a while, my solution for qualified expressions was to avoid a block with the expression indent, and the reason was to allow the optimization for a final lambda without extra indentation as in:
```
foo.blahblah.forEach {
  // only 2 spaces indent
}
```

This kept causing problems. I now realize I had it wrong. I should put in the expression indent block, but close it before the last expression if it is a lambda. This meant redoing a lot of the function, but it should be cleaner and better to read now.

Reviewed By: cgrushko

Differential Revision: D21428681

fbshipit-source-id: 9df02a2501d5fadc85f41a7c87df2072cadb144b