Drop redundant semicolons
Summary:
dropRedundantElements is a hack that tries to detect a necessary semicolon; it would be much better to use https://github.com/Kotlin/grammar-tools, which would return `SEMICOLON` for necessary semicolons, and `semi` for optional ones. However, it's marked experimental, and the README says it could sometimes diverge from kotlin-compiler.

So for now, I went with looking at ASTs of necessary semicolons and manually checking if that's the case in the code as well.

I used the grammar definition (https://github.com/Kotlin/kotlin-spec/blob/master/grammar/src/main/antlr/KotlinParser.g4) to code the rules for a semicolon to be necessary.

This fixes https://github.com/facebookincubator/ktfmt/issues/38

Reviewed By: strulovich

Differential Revision: D21544559

fbshipit-source-id: 2b40be467dbcba833284859983d60c2afde0b6e7