Do not break due to annotations parameters
Summary:
This is basically going counter to the advice in the official guidelines:

https://kotlinlang.org/docs/reference/coding-conventions.html#annotation-formatting

The difference is apparent in tests. I think the official guidelines don't have enough detail and granularity. They want annotations generally on their own line, which is what we will do whenever the thing they annotate is not a 1 liner (so most classes do have a line break per annotations since they're likely to be more than one line)

This was becoming noticeable in our code using the open source Litho where a lot of parameters are annotated with `Prop(optional = true)`. Breaking due to those was making the code harder to read.

Reviewed By: cgrushko

Differential Revision: D24478713

fbshipit-source-id: b769df6ea10dc363ae7cb8a97e2c528253c05f0e