Temp: [Ktfmt] Don't break before lambda when initializing property
Summary:
See test.

Before this we would break according to rectangle rule for lambdas so the code would be:
```
val lambda =
    {
      ...
    }
```

The result obeys the rectangle rule, but people naturally format Kotlin the other way, and we already practice a similar mode for named arguments. So this unifies: `that = { ... }` when `that` is a property with when it's a named argument to function.

Reviewed By: hick209

Differential Revision: D25283008

fbshipit-source-id: 6d9525c86ab4ec9b5be41eed27b285a72f1d4b4b