Updating primary constructor indentation
Summary:
According to the [coding guidelines](https://kotlinlang.org/docs/reference/coding-conventions.html#class-header-formatting) in the Kotlin website, the goal is to have properties declared in the primary constructor at the same level of the ones declared in the body of the class.

> Rationale: This ensures that properties declared in the primary constructor have the same indentation as properties declared in the body of a class.

There are no explicit examples of this use case, but this format does look appealing to me and I would like to a least start a discussion whether or not we should adopt this.

BTW, this is the format that Portal currently adopts.

Reviewed By: strulovich

Differential Revision: D25320113

fbshipit-source-id: 9266f7171e6b1400c620a618f1ddbe97a781f56a