Output expected formatting when showing a mismatched formatting error
Summary: When `assertThatFormatting(code).isEqualTo(foo)` fails, it shows the input and output, but not what we expected which forces you to look at the test case to understand why the output is incorrect. This adds the expected output so you dont necessarily have to look at the test case.

Reviewed By: cgrushko

Differential Revision: D26628768

fbshipit-source-id: c90685f058b768980f3706f31c1fe72551d6ec90