Add Spotless (GoogleJavaFormat) to IDEA plugin (#184)
Summary:
To ensure the IDEA plugin code is formatted correctly, Spotless will check it every time a push or PR.

I tweaked the plugin workflow created by jef to just use `./gradlew build` which depends on the task `spotlessJavaCheck`, so it will check if the plugin code compiles, tests pass, and if the code is correctly formatted.

To automatically format the code, run `./gradlew spotlessJavaApply`.

Pull Request resolved: https://github.com/facebookincubator/ktfmt/pull/184

Reviewed By: hick209

Differential Revision: D26714553

Pulled By: cgrushko

fbshipit-source-id: 47d499a96e4f30bc314033897de1e1d7a5977403