Fix formatting of block-like lambdas (+ scope functions) on long lines. (#280)
Summary:
If everything up to and including the `->` can fit on the current line, then indent only the lambda body. If we must break before the `->`, treat the lambda and scoping function as an expression, and indent the entire thing into a rectangle.

Pull Request resolved: https://github.com/facebookincubator/ktfmt/pull/280

Test Plan: ktfmt-diff: D34240625

Reviewed By: strulovich

Differential Revision: D34199844

Pulled By: cgrushko

fbshipit-source-id: 6cac4bd2817952140f7982f15545a0d4bfb42882