Proof of concept for propagating breaks along call expression blocklikes.
There's a bad interaction between this change and the current implementation of fluent chain
formatting, however it should give a feel for the effect. I didn't want to spend too much time
polishing this until we had the general idea worked out. It would probably also be smoother
to rework fluent chains first before working on this change.