Always preserve semicolons preceeding lambda expressions, even when they might be meaningless.
There are a huge number of cases where removal might change behaviour, because the trailing
lambda syntax is so flexible. Therefore, we just assume that all semicolons followed by
lambdas are meaningful. The cases where they could be removed are too rare to justify the
risk of changing behaviour.

https://github.com/facebookincubator/ktfmt/pull/278 tried to fix this earlier, but missed
a lot of cases. Some examples are added to the tests.