Fix memory leak when creating KtFile
Summary:
The disposable is leaking, and according to this it's a known issue:
https://github.com/JetBrains/kotlin/blob/master/compiler/cli/src/org/jetbrains/kotlin/cli/jvm/compiler/KotlinCoreEnvironment.kt#L544

I could only find one thing through all the chaining:
https://github.com/JetBrains/kotlin/blob/master/compiler/cli/src/org/jetbrains/kotlin/cli/jvm/compiler/KotlinCoreEnvironment.kt#L149 -> clearing some cache

So instead of creating this environment per each file, causing a memory leak, we make one singleton item and ride it all the way to the end of the binary. While this item may grow in size, it seems to now do so, definitely not as the current memory leak.

See results of how faster the code is running on many files.

Reviewed By: hick209

Differential Revision: D35192326

fbshipit-source-id: 5308555bd689c18214ca2b824d28b8c283586572