Redo qualified chains, without chunks, other modifications and support handling !! and []
Summary:
Ok, davidtorosyan and I chatted and I learned that none of us understands emitQualifiedExpression, so we're doomed to rewrite it every time a change is needed.

This is an attempt to simplify it a bit - which doesn't mean much. So I will spend the time cleaning and making this more readable later and sit with you all to try and avoid this in the future.

What do we have here?
- Eliminate usage of chunks
- To handle qualified expressions look at the chain parts, then mark where we need to open and close a group.
- Then "just" emit stuff using this metadata
- We now traverse array expressions (i.e. `[1, 2]`) and unsafe dereferences (i.e. `!!`) as part of the chain. This will improve rendering of them. (Note that `as`, `::` and maybe others may need similar treatment, punting on this

Many tests have been updated, to neutral changes (i.e. I have no opinion) but some for the better.

Reviewed By: cgrushko

Differential Revision: D35224180

fbshipit-source-id: 8de4c633ede5f26d4712d5fb5c9d9b0a32454ee9