For annotated expressions, break after annotations iff it is block-level expression.
For some block-level expressions (e.g. binary operators, infix functions) the line break changes the target of the annotation. When on the previous line, the annotation targets the entire expression. When on the same line, it targets only the left-most sub-expression.

In order to always preserve semantics, and guarantee consistent formatting, we always break only for this case.