Add line breaks to lambdas after broken function arguments
Summary:
We've had some requests to format this:

```
foo.bar(
  trailingComma,
) { x = 0 }
```

As this:
```
foo.bar(
  trailingComma,
) {
  x = 0
}
```

Here I do that by keeping track of when we enter lambda arguments in call elements and noting if there's a trailing comma.

## Caveat

**Google-style requires a trailing comma**

The Google-style formatting puts the closing paren on the next line, even when there's no trailing comma. Due to this we can't detect that we need to put a break in the lambda.

This could be fixed by detecting if the params broke (instead of looking for a trailing comma) but I'm not sure how to do that.

Reviewed By: strulovich

Differential Revision: D36649191

fbshipit-source-id: 2c8d316f068c1030437af59b5114ffae9d5d2aa8