Simplify visitValueArgumentList, unite two logic branches
Summary:
WARNING: This is NOT purely a refactor, there's a slight change.

The `visitValueArgumentList` function handles single unnamed lambdas differently from other arguments, which makes it hard to change the behavior of. This unifies the two branches.

Note that this has a side effect of making our formatting a bit more consistent (see the changed unit test).

Reviewed By: hick209

Differential Revision: D37398560

fbshipit-source-id: 5cd3360d2d616558ce70c09e78d8b1bc41b6d9d6