Simplify visitTypeArgumentList, move bracket logic into helper
Summary:
WARNING: This is NOT purely a refactor, there's a slight change.

This change consolidates list logic into `visitEachCommaSeparated`. The next diff will make a similar change to `visitValueArgumentList` so we can change the logic of both in one go.

Note that this does make a small change: the type argument list now works the same as the value argument list, so in Google style a break will put the closing bracket on the next line.

Reviewed By: strulovich

Differential Revision: D37399796

fbshipit-source-id: 575fdd5c37dd80ed62c614f46aef44e6ca5401ca