Fix some spacing issues with comma separated lists in --google-style
After the recent improvments to comma separated list handling, there
are a few cases that suffered collateral damage under --google-style.
This PR is a quick fix to get those cases back to how they used to be,
and add some regression tests. It doesn't necessarily represent the
ideal --gogole-style output.