Only pretty-print each file one time
The previous formatter flow pretty-printed each file twice, in
order not to change the meaning of the code when removing
semicolons. This made debugging more difficult, because the
second pretty-print was running on already transformed code,
and also caused strange crashes if the result of the first
pretty-print was malformed.

This change eliminates the first pretty-print, running it only
once as the final formatting step. This is done by swapping in
newline charaacters, where appropriate, during semicolon removal.

A side-effect of this change is that a spacing bug is revealed
in the rare case of `if (cond) /** doc */ ;`. Apparently double
pretty-printing hid this behaviour. In exchange, a related bug
is fixed: `if (cond) ; else x` no longer crashes.