Fix unneeded break following long multiline string (google style) (#348)
Summary:
## Issue

Fix this issue filed on GitHub:
* [trailing function call parenthesis regression in 0.40 · Issue #348 · facebookincubator/ktfmt · GitHub](https://github.com/facebookincubator/ktfmt/issues/348)

See new unit tests added in D39289539.

## Context

D37401739 (https://github.com/facebookincubator/ktfmt/commit/c32319170e4771d377dee27eef1d4bbeb1ca0ef0) was supposed to be a no-impact refactor, but it put the trailing break in arguments outside of the block with the parens. This only affected Google style, since otherwise the arguments are wrapped in another block anyway.

## Fix

To fix, just check for empty arguments before applying the trailing break.

To demonstrate that the fix is correct, I applied the unit tests to the version just before the regressed code (see D39293065). You can see that the behavior is back to how it used to be.

Closes #348

Reviewed By: cgrushko

Differential Revision: D39293043

fbshipit-source-id: 23e731d247f91e7f00c548235f7e48bd64d6f30b