Only pretty-print each file one time (#345)
Summary:
The previous formatter flow pretty-printed each file twice, in
order not to change the meaning of the code when removing
semicolons. This made debugging more difficult, because the
second pretty-print was running on already transformed code,
and also caused strange crashes if the result of the first
pretty-print was malformed.

This change eliminates the first pretty-print, running it only
once as the final formatting step. This is done by swapping in
newline charaacters, where appropriate, during semicolon removal.

A side-effect of this change is that a spacing bug is revealed
in the rare case of `if (cond) /** doc */ ;`. Apparently double
pretty-printing hid this behaviour. In exchange, a related bug
is fixed: `if (cond) ; else x` no longer crashes.

Pull Request resolved: https://github.com/facebookincubator/ktfmt/pull/345

Reviewed By: davidtorosyan

Differential Revision: D39473841

Pulled By: cgrushko

fbshipit-source-id: 8103d87696ae5431c031e3ab80e843f931026849