Avoid early break at TODO:
Summary:
As discussed in this kdocformatter issue [Early break caused by @](https://github.com/tnorbye/kdoc-formatter/issues/64), it's not a good idea to break before meaningful tokens.

Well, `TODO:` is a meaningful token, since it gets moved to the end of a kdoc. In this diff, add it to the list of tokens you can't break at.

Reviewed By: cgrushko

Differential Revision: D41454151

fbshipit-source-id: a71eb9b7c3623a173398d3101708076196b716fa