Eliminate mis-paired OpsBuilder.open()/close() (#379)
Summary:
These explicit calls are misleading because they pair with the calls inside OpsBuilder.block(), so the indention of the source doesn't match the block structure of the output.

I believe the only potentially affected case is unnamed function-literal-type params, which cannot have initializers, making this a no-op.

Pull Request resolved: https://github.com/facebook/ktfmt/pull/379

Reviewed By: strulovich

Differential Revision: D42768487

Pulled By: cgrushko

fbshipit-source-id: caed350dd7c78ee33df29c60b60cdf013bb9b1e8