Add unit tests to capture line break behavior on type specifiers
Summary:
It came up recently that this formatting (for the `where` case) looks weird and should probably be indented.

To make sure we track any fixes (or just are aware that we've always had this issue) add a couple of unit tests. One to demonstrate the issue, the other to show a better case.

Note that a previous change (D34347730) added a similar (but different) unit test.

Reviewed By: hick209

Differential Revision: D49507677

fbshipit-source-id: 777eb3becf51d0061fea19953fd950a71d8fe36b