Pointing Gradle deps to use latest stable version of ktfmt
Summary:
Right now things are broken as is since we don't update snapshot versions of ktfmt to maven.

My proposal is to use the latest available stable version when compiling them.
Likely not optimal, but at least should reduce the curernt amount of headaches.

Issues:
```
FAILURE: Build failed with an exception.

* What went wrong:
Execution failed for task ':compileKotlin'.
> Could not resolve all files for configuration ':compileClasspath'.
   > Could not find com.facebook:ktfmt:0.49-SNAPSHOT.
```
and
```
FAILURE: Build failed with an exception.

* What went wrong:
Execution failed for task ':compileJava'.
> Could not resolve all files for configuration ':compileClasspath'.
   > Could not find com.facebook:ktfmt:0.49.
```

Reviewed By: cortinico

Differential Revision: D56472377

fbshipit-source-id: 509620821be40353c10e41f72670de5567792915