Convert IntelliJ plugin to Kotlin (#502)
Summary:
> [!NOTE]
> This requires https://github.com/facebook/ktfmt/issues/501 to be merged first

This PR is the basis for future improvements of the plugin. In this PR, the plugin code is converted to Kotlin, the configurable is recreated with the Kotlin UI DSL, and the goal is to maintain the existing behaviours.

The only small change is that the style dropdown in the configurable is disabled when ktfmt is disabled.

Pull Request resolved: https://github.com/facebook/ktfmt/pull/502

Reviewed By: strulovich

Differential Revision: D60968984

Pulled By: hick209

fbshipit-source-id: 467ce14d21d06c783d8e9254e913f966dcbff8ca