fix: comment max width (#511)
Summary:
`FormattingOptions#maxWidth` is not respected for double slash comments. This is because it wasn't using the supplied value, and instead checked against the static property `com.google.googlejavaformat.java.Formatter.MAX_LINE_LENGTH`. I was unable to find any documentation / discussion that would indicate this as being intentional.

Test cases that deduced `maxWidth` actually relied on this behavior in some cases. Some of the fixes in 84dbfae449866787ba088d5ea9f0bad48b626cf5 simply adjusted the number of `/`, while others relied on a specific `maxWidth` and so I manually line-broke the comments to preserve the behavior under test.

Pull Request resolved: https://github.com/facebook/ktfmt/pull/511

Reviewed By: strulovich

Differential Revision: D61711870

Pulled By: hick209

fbshipit-source-id: d27479c17f68d376c14722a8d3997ea602cfe723