Support context parameters (#536)
Summary:
Resolves https://github.com/facebook/ktfmt/issues/518

Note that in order for `LightVirtualFile` to correctly parse this, it did require updating to the Kotlin 2.2.0 betas. Since this ships as a fat jar I think it would be fine to build against this in `main`, but up to y'all!

This adds a couple new tests for it + ensures backward compatibility with the previous context receivers syntax.

Pull Request resolved: https://github.com/facebook/ktfmt/pull/536

Reviewed By: cortinico, strulovich

Differential Revision: D74587694

Pulled By: hick209

fbshipit-source-id: 2250209265372a74f55a0f1a3767d0df25c28171