Fix gradle scripted builds
Summary:
Issue:
https://github.com/facebook/ktfmt/actions/runs/15147530441/job/42586736383#step:4:28

Repro:
```
cd xplat/ktfmt
./gradlew clean
./gradlew :ktfmt:build
```

Fix:
```
./gradlew :ktfmt:compileKotlin
# now this will pass too
./gradlew :ktfmt:build
```

It turns that if we compile any of the other projects before the main one, then it compiles fine afterwards, which I guess means that ktfmt-gradle was compiled and is now available.

I don't fully understand what is going on here, but I'm glad this will help us stop the bleeding.

Differential Revision: D75134683

fbshipit-source-id: 1ec57fab4f49d130e591d31363caac7358699eaa