Fixed release workflow
Summary:
I finally set things up locally to test workflows.
It was already worth the effort as things did not work out on the first try. In fact it took 6 iterations, therefore it would have taken a few tries to validate that it was working, which would have been more time consuming for sure.

Reviewed By: cortinico

Differential Revision: D75220352

fbshipit-source-id: 91cba21603c0220d3650300239f58f661cd22dbf