Adjust nexus publishing to auto promote staged releases
Summary:
We have finally been able to successfully publish version 0.55, but it is now in "staging" mode and not publicly available in maven central yet, waiting for promotion, which could be done manually.

Instead of doing that everytime, I've configured things to automatically promote the releases.

I've used the plugin from here to do this:
https://github.com/gradle-nexus/publish-plugin

Differential Revision: D75287203

fbshipit-source-id: 7bffdb7c4ad957ea147f14d0e25da71a0bf8f64b