Do not to add extra line on the beginning
Summary:
Fixed a couple of issues + restructured the code to make it more maintainable. Adding more tests to it as well, which were the result of the work to make it more modular.

Issues fixed:
1. Fixed issue where formatter would remove nested multiline triple quotes
2. Do not to add extra line on the beginning

## Issue 1 : formatter removes nested multiline triple quotes
Before, blocks such as this
```
"""
${
  """
  hello
  """
      .trimIndent()
}
"""
    .trimIndent()
```
Would be turned into this here, which breaks compilation
```
"""
${
  """
  hello
      .trimIndent()
}
"""
    .trimIndent()
```

This is an unacceptable bug! This here fixes it, but I've also disabled formatting of nested trim indent since formatting template expressions is not something we do in ktfmt at the moment.

## Issue 2 : Do not to add extra line on the beginning
I've validated and this is what Kotlin does using the following code
```
print("<output>")
print("""${"    "}
    hello
    world
    """.trimIndent())
println("</output>")

println("---")

print("<output>")
print("""${"    "}
    hello
    world
    """.trimMargin())
println("</output>")
```
Results in
```
<output>hello
world</output>
 ---
<output>    hello
    world</output>
```

Therefore here I make sure we don't include the extra line at the begining as we currently do.

Differential Revision: D81396128

fbshipit-source-id: 388a0e7b1869c377be86481a135a6615f62438b7