fix: pass absolute path to editorconfig properties service
queryProperties walks up the provided path to find `.editorconfig`
files.

In the common case of a `.editorconfig` in the root of the repo or
project, when running `ktfmt` on a path like `src/main/kotlin/MyFile.kt`
it will stop looking in `src` as `src` has no "parent" and not find any
`.editorconfig`. Using a path like `./src/main/kotlin/MyFile.kt` will
stop at `.` and find `./.editorconfig`.

By using the absolute path of the file, it'll walk all ancestors until
it finds a `.editorconfig` with `root = true` as expected.